% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fn.R
\name{build_fnn_model}
\alias{build_fnn_model}
\title{Build and compile the FNN model}
\usage{
build_fnn_model(input_shape)
}
\arguments{
\item{input_shape}{Number of input features}
}
\value{
A compiled Keras model
}
\description{
Build and compile the FNN model
}
