% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate model on new data}
\usage{
evaluate_model(model_file, X, y, weights)
}
\arguments{
\item{model_file}{File path of fine-tuned model}

\item{X}{Matrix of input features}

\item{y}{True labels}

\item{weights}{Sample weights}
}
\description{
Evaluate model on new data
}
