% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finetune_model.R
\name{finetune_model}
\alias{finetune_model}
\title{Fine-tune a pretrained STL model}
\usage{
finetune_model(model_file, X_target, y_target, freeze_layers = TRUE)
}
\arguments{
\item{model_file}{Path to the pretrained model file}

\item{X_target}{Feature matrix of the target dataset}

\item{y_target}{Outcome vector of the target dataset}

\item{freeze_layers}{Logical, whether to freeze the early layers}
}
\value{
A fine-tuned Keras model
}
\description{
Fine-tune a pretrained STL model
}
