% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess data for STL model}
\usage{
preprocess_data(df, outcome, features)
}
\arguments{
\item{df}{A data frame with features and outcome}

\item{outcome}{Column name of outcome}

\item{features}{Vector of predictor column names}
}
\value{
A list with processed X, y
}
\description{
Preprocess data for STL model
}
