% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stl.R
\name{read_stl}
\alias{read_stl}
\title{Read an STL file}
\usage{
read_stl(path, verbose = FALSE)
}
\arguments{
\item{path}{\link{character} (\strong{required}): Path to a stored stl file}

\item{verbose}{\link{logical} (\strong{required}): Provide further information about the parsed file}
}
\value{
A matrix
}
\description{
Read an STL file
}
