% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stl-package.R
\docType{package}
\name{stl-package}
\alias{stl}
\alias{stl-package}
\title{stl: Survey Transfer Learning for Generating Synthetic Responses}
\description{
Provides tools to apply deep learning models to transfer knowledge between national surveys such as CES and ANES. Enables prediction of missing variables or unasked questions using shared demographic features. Implements a feedforward neural network model with optional fine-tuning on target datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aliiamini/stl}
  \item Report bugs at \url{https://github.com/aliiamini/stl/issues}
}

}
\author{
\strong{Maintainer}: Ali Amini \email{aa6718a@american.edu}

}
\keyword{internal}
