% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_fnn_model.R
\name{train_fnn_model}
\alias{train_fnn_model}
\title{Train the STL model}
\usage{
train_fnn_model(
  model,
  X_train,
  y_train,
  weights = NULL,
  filename = "ces_pretrained_model.h5"
)
}
\arguments{
\item{model}{A compiled Keras model}

\item{X_train}{Feature matrix}

\item{y_train}{Label vector}

\item{weights}{Sample weights (optional)}

\item{filename}{Where to save the model}
}
\description{
Train the STL model
}
